+++ ChangeLog.md
@@ -14,6 +14,7 @@
   * Ensure that personality names do not contain illegal file or path characters to avoid issues when saving them
   * Send EDDN messages with "Unknown Commander" when commander name is not known
   * Use recursive/dynamic method to populate VoiceAttack variables.  This provides many more VoiceAttack variables than were previously available; details are in the relevent documentation
+  * Add VoiceAttack 'disablespeechresponder' and 'enablespeechresponder' contexts.  Details on these are in the VoiceAttack documentation.
 
 #2.0.13
   * Fix issue where engineer rank journal message without rank would cause a crash